* commit '1df823e3e3bf8cfce7d0912fd41e755809f2f5ca':
  Revert "Convert jint to jlong for native pointers in GLFunctionTable to support 64 bit."

+++ chromium/java/com/android/webview/chromium/GraphicsUtils.java
@@ -17,15 +17,15 @@
 package com.android.webview.chromium;
 
 abstract class GraphicsUtils {
-    public static long getDrawSWFunctionTable() {
+    public static int getDrawSWFunctionTable() {
         return nativeGetDrawSWFunctionTable();
     }
 
-    public static long getDrawGLFunctionTable() {
+    public static int getDrawGLFunctionTable() {
         return nativeGetDrawGLFunctionTable();
     }
 
-    private static native long nativeGetDrawSWFunctionTable();
-    private static native long nativeGetDrawGLFunctionTable();
+    private static native int nativeGetDrawSWFunctionTable();
+    private static native int nativeGetDrawGLFunctionTable();
 
 }
